requirements.txt
MetaTrader5
pandas
pytz
streamlit
TA-Lib
openai
pyyaml
python-dotenv


